Removed Check_key and added some functions to input_data_functions.py
