+++ tests/functional-tests/ibm_spectrum_scale_csi/scale_operator.py
@@ -9,6 +9,7 @@ import ibm_spectrum_scale_csi.kubernetes_apis.volume_functions as volfunc
 import ibm_spectrum_scale_csi.kubernetes_apis.snapshot_functions as snapshotfunc
 import ibm_spectrum_scale_csi.spectrum_scale_apis.fileset_functions as filesetfunc
 import ibm_spectrum_scale_csi.kubernetes_apis.cleanup_functions as cleanup
+import ibm_spectrum_scale_csi.common_utils.input_data_functions as inputfunc
 LOGGER = logging.getLogger()
